Refactor files by tapis service
