+++ client/src/components/index.js
@@ -1,5 +1,2 @@
-import App from './App';
-
-// To be replaced later with:
-// export { default as App } from './App';
-export default App;
+/* eslint-disable */
+export { default as Login } from './Login';
