+++ client/src/tapis-redux/index.js
@@ -1,1 +1,2 @@
-export { default as useLogin } from './hooks/useLogin';
+/* eslint-disable */
+export { default as useLogin } from './auth/useLogin';
