+++ client/src/tapis-redux/reducers/index.js
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux';
-import auth from './auth.reducer';
+import auth from '../auth/auth.reducer';
 
 export default combineReducers({
   auth,
