+++ client/src/tapis-redux/sagas/index.js
@@ -1,5 +1,5 @@
 import { all } from 'redux-saga/effects';
-import { watchLogin } from './auth';
+import { watchLogin } from '../auth/auth.saga';
 
 export default function* rootSaga() {
   yield all([watchLogin()]);
